Move/Change Folder Structure, Add Login Changes
Added a new login system that only uses accounts defined in users.json.
Each account has a purpose and you can use a method defined in myvwlogin to
load account details from the file that match a given purpose.

Also adjusted the folder structure of lib to match more what is in the features
folder. Some files were moved into Specail_pages, such as VDT, Dogs etc
This is to improve consistancy and make it easier to find files